requirements.txt
django==4.0.10
djongo==1.3.6
pymongo==3.13.0
sqlparse==0.2.4
pytz
dnspython
python-gettext
python-dotenv
Pillow
whitenoise
opencv-python-headless==4.7.0.72
daphne
channels==3.0.5
channels_redis
django-allauth
openai
asgiref
requests